fix: final stages

+++ shared/pkg/messenger/models/threshold_update.go
@@ -5,5 +5,5 @@ import "zhacked.me/oxyl/shared/pkg/models"
 type ThresholdUpdate struct {
 	CompanyId     string                  `json:"company_id"`
 	ThresholdType models.NotificationType `json:"notification_type"`
-	Threshold     int                     `json:"threshold"`
+	Threshold     float64                 `json:"threshold"`
 }
